Move `croud organizations` tests to dedicated module
