+++ tests/test_commands.py
@@ -20,12 +20,7 @@
 import uuid
 from unittest import mock
 
-import pytest
-
 from croud.config import Configuration
-from croud.organizations.users.commands import (
-    role_fqn_transform as organization_role_fqn_transform,
-)
 from croud.projects.users.commands import (
     role_fqn_transform as project_role_fqn_transform,
 )
@@ -37,346 +32,6 @@ from tests.util import CommandTestCase, call_command
 
 def gen_uuid() -> str:
     return str(uuid.uuid4())
-
-
-config_org_id = gen_uuid()
-FALLBACK_ORG_ID_CONFIG: dict = {
-    "auth": {
-        "current_context": "prod",
-        "contexts": {
-            "prod": {"token": "", "organization_id": config_org_id},
-            "dev": {"token": "", "organization_id": config_org_id},
-            "local": {"token": "", "organization_id": config_org_id},
-        },
-    },
-    "region": "bregenz.a1",
-    "output_fmt": "table",
-}
-
-
-class TestOrganizations(CommandTestCase):
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_create(self, mock_send, mock_load_config):
-        argv = ["croud", "organizations", "create", "--name", "test-org"]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.POST,
-            "/api/v2/organizations/",
-            body={"name": "test-org"},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_create_name_plan_type(self, mock_send, mock_load_config):
-        argv = [
-            "croud",
-            "organizations",
-            "create",
-            "--name",
-            "test-org",
-            "--plan-type",
-            "3",
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.POST,
-            "/api/v2/organizations/",
-            body={"name": "test-org", "plan_type": 3},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_edit_name(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-        argv = [
-            "croud",
-            "organizations",
-            "edit",
-            "--name",
-            "new-org-name",
-            "--org-id",
-            org_id,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.PUT,
-            f"/api/v2/organizations/{org_id}/",
-            body={"name": "new-org-name"},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_edit_plan_type(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-        argv = [
-            "croud",
-            "organizations",
-            "edit",
-            "--plan-type",
-            "3",
-            "--org-id",
-            org_id,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.PUT,
-            f"/api/v2/organizations/{org_id}/",
-            body={"plan_type": 3},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_edit_name_plan_type(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-        argv = [
-            "croud",
-            "organizations",
-            "edit",
-            "--name",
-            "new-org-name",
-            "--plan-type",
-            "3",
-            "--org-id",
-            org_id,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.PUT,
-            f"/api/v2/organizations/{org_id}/",
-            body={"name": "new-org-name", "plan_type": 3},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=(None, {}))
-    def test_edit_no_arguments(self, mock_send, mock_load_config, capsys):
-        org_id = gen_uuid()
-        argv = ["croud", "organizations", "edit", "--org-id", org_id]
-        with pytest.raises(SystemExit):
-            self.assertRest(
-                mock_send, argv, RequestMethod.PUT, f"/api/v2/organizations/{org_id}/"
-            )
-        _, err = capsys.readouterr()
-        assert "No input arguments found." in err
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_list(self, mock_send, mock_load_config):
-        argv = ["croud", "organizations", "list"]
-        self.assertRest(mock_send, argv, RequestMethod.GET, "/api/v2/organizations/")
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=(None, {}))
-    def test_delete(self, mock_send, mock_load_config, capsys):
-        org_id = gen_uuid()
-        argv = ["croud", "organizations", "delete", "--org-id", org_id]
-        with mock.patch("builtins.input", side_effect=["Y"]) as mock_input:
-            self.assertRest(
-                mock_send,
-                argv,
-                RequestMethod.DELETE,
-                f"/api/v2/organizations/{org_id}/",
-            )
-            mock_input.assert_called_once_with(
-                "Are you sure you want to delete the organization? [yN] "
-            )
-
-        _, err_output = capsys.readouterr()
-        assert "Organization deleted." in err_output
-
-    @mock.patch("croud.config.Configuration.set_organization_id")
-    @mock.patch("croud.config.load_config", return_value=FALLBACK_ORG_ID_CONFIG)
-    @mock.patch.object(Client, "send", return_value=(None, {"errors": "Message"}))
-    def test_delete_failure_org_id_not_deleted_from_config(
-        self, mock_send, mock_load_config, mock_set_config, capsys
-    ):
-        argv = ["croud", "organizations", "delete"]
-        with mock.patch("builtins.input", side_effect=["Y"]):
-            self.assertRest(
-                mock_send,
-                argv,
-                RequestMethod.DELETE,
-                f"/api/v2/organizations/{mock_load_config.return_value['auth']['contexts']['local']['organization_id']}/",  # noqa
-            )
-        _, err = capsys.readouterr()
-        assert err is not None
-        assert (
-            mock_set_config.called is False
-        )  # This means that the org_id is still in the local config
-
-    @mock.patch("croud.config.Configuration.set_organization_id")
-    @mock.patch("croud.config.load_config", return_value=FALLBACK_ORG_ID_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_delete_org_id_from_local_config(
-        self, mock_send, mock_load_config, mock_set_config, capsys
-    ):
-        argv = ["croud", "organizations", "delete"]
-        with mock.patch("builtins.input", side_effect=["Y"]):
-            self.assertRest(
-                mock_send,
-                argv,
-                RequestMethod.DELETE,
-                f"/api/v2/organizations/{mock_load_config.return_value['auth']['contexts']['local']['organization_id']}/",  # noqa
-            )
-        _, err = capsys.readouterr()
-        assert (
-            mock_set_config.called is True
-        )  # This means that the org_id was deleted from the local config
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=(None, {}))
-    def test_delete_flag(self, mock_send, mock_load_config, capsys):
-        org_id = gen_uuid()
-        argv = ["croud", "organizations", "delete", "--org-id", org_id, "-y"]
-        with mock.patch("builtins.input") as mock_input:
-            self.assertRest(
-                mock_send,
-                argv,
-                RequestMethod.DELETE,
-                f"/api/v2/organizations/{org_id}/",
-            )
-            mock_input.assert_not_called()
-
-        _, err_output = capsys.readouterr()
-        assert "Organization deleted." in err_output
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @pytest.mark.parametrize("input", ["", "N", "No", "cancel"])
-    @mock.patch.object(Client, "send", return_value=(None, {}))
-    def test_delete_aborted(self, mock_send, mock_load_config, capsys, input):
-        org_id = gen_uuid()
-        argv = ["croud", "organizations", "delete", "--org-id", org_id]
-        with mock.patch("builtins.input", side_effect=[input]) as mock_input:
-            call_command(*argv)
-            mock_send.assert_not_called()
-            mock_input.assert_called_once_with(
-                "Are you sure you want to delete the organization? [yN] "
-            )
-
-        _, err_output = capsys.readouterr()
-        assert "Organization deletion cancelled." in err_output
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=(None, {}))
-    def test_delete_aborted_with_input(self, mock_send, mock_load_config, capsys):
-        org_id = gen_uuid()
-        argv = ["croud", "organizations", "delete", "--org-id", org_id]
-        with mock.patch("builtins.input", side_effect=["N"]) as mock_input:
-            call_command(*argv)
-            mock_send.assert_not_called()
-            mock_input.assert_called_once_with(
-                "Are you sure you want to delete the organization? [yN] "
-            )
-
-        _, err_output = capsys.readouterr()
-        assert "Organization deletion cancelled." in err_output
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({"added": True}, None))
-    def test_add_user(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-        user = "test@crate.io"
-        role_fqn = "org_admin"
-
-        argv = [
-            "croud",
-            "organizations",
-            "users",
-            "add",
-            "--user",
-            user,
-            "--org-id",
-            org_id,
-            "--role",
-            role_fqn,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.POST,
-            f"/api/v2/organizations/{org_id}/users/",
-            body={"user": user, "role_fqn": role_fqn},
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({"added": False}, None))
-    def test_update_user(self, mock_send, mock_load_confg, capsys):
-        org_id = gen_uuid()
-        user = "test@crate.io"
-        role_fqn = "org_admin"
-
-        argv = [
-            "croud",
-            "organizations",
-            "users",
-            "add",
-            "--user",
-            user,
-            "--org-id",
-            org_id,
-            "--role",
-            role_fqn,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.POST,
-            f"/api/v2/organizations/{org_id}/users/",
-            body={"user": user, "role_fqn": role_fqn},
-        )
-        _, err_output = capsys.readouterr()
-        assert "Success" in err_output
-        assert "Role altered for user." in err_output
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    def test_role_fqn_transform(self, mock_load_config):
-        user = {
-            "organization_roles": [
-                {"organization_id": "org-1", "role_fqn": "organization_admin"},
-                {"organization_id": "org-2", "role_fqn": "organization_member"},
-                {"organization_id": "org-3", "role_fqn": "organization_member"},
-            ]
-        }
-        response = organization_role_fqn_transform(user["organization_roles"])
-        assert response == "organization_admin"
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_list_user(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-
-        argv = ["croud", "organizations", "users", "list", "--org-id", org_id]
-        self.assertRest(
-            mock_send, argv, RequestMethod.GET, f"/api/v2/organizations/{org_id}/users/"
-        )
-
-    @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
-    @mock.patch.object(Client, "send", return_value=({}, None))
-    def test_remove_user(self, mock_send, mock_load_config):
-        org_id = gen_uuid()
-        user = "test@crate.io"
-
-        argv = [
-            "croud",
-            "organizations",
-            "users",
-            "remove",
-            "--user",
-            user,
-            "--org-id",
-            org_id,
-        ]
-        self.assertRest(
-            mock_send,
-            argv,
-            RequestMethod.DELETE,
-            f"/api/v2/organizations/{org_id}/users/{user}/",
-        )
 
 
 @mock.patch("croud.config.load_config", return_value=Configuration.DEFAULT_CONFIG)
